+++ packages/types/wcag-techniques.d.ts
@@ -1,5 +1,4 @@
 declare module "@qualweb/wcag-techniques" {
-  import { Position } from "css";
   import { QWPage } from "@qualweb/qw-page";
 
   interface WCAGOptions {
@@ -61,7 +60,6 @@ declare module "@qualweb/wcag-techniques" {
       value?: string;
     };
     stylesheetFile?: string;
-    position?: Position;
   }
 
   interface WCAGTechnique {